feat(rgb): use SPI via PIO for RGB underglow
ZMK currently supports using SPI (via PIO as well, if required) for
RGB underglow.[^1]

ZMK is on Zephyr 3.5 currently[^2], and doesn't include the patches that
add PIO-based WS2812 driver for `rpi_pico`, which are available since
Zephyr 3.6 only.[^3][^4]

So, update the board definition to use SPI via PIO to run the RGB
underglow. Once ZMK itself updates to a Zephyr version that includes the
requisite patches, this can be reverted to use the PIO-based driver
directly.

[^1]: https://zmk.dev/docs/development/hardware-integration/lighting/underglow#other-boards
[^2]: https://github.com/zmkfirmware/zmk/blob/cef7af4408cc44c20fab93a0b2e20b3429d0a98e/app/west.yml#L8-L10
[^3]: https://github.com/zephyrproject-rtos/zephyr/commit/6699d4d4f931f5f686bae65da5f89589395d86ba
[^4]: https://github.com/zephyrproject-rtos/zephyr/commit/0f458c9564d76e9e077230dc321cbddb1e801cc9